requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
neo4j==5.14.1
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.6
pydantic-settings==2.1.0
groq==0.4.2
